Crate slk581

The crate provides function to encode data into SLK581 format.

The format allows encode given family name, given name, date of birth and sex into sequence XXXZZDDMMYYYYN. Where XXX encodes family name, ZZ encodes given name, DDMMYYYY encodes date of birth and N encodes sex.

Enums

SLK581Error

Constants

FEMALE
Female code 2

INPUT_DATE_FORMAT
Supported input format of date of birth YYYY-MM-DD

MALE
Male code 1

OUTPUT_DATE_FORMAT
Output format of date of birth DDMMYYYY

TRANSGENDER
Transgender code 3

UNKNOWN_CHARACTER_IN_NAME
Placeholder for missing character in given or family name 2

UNKNOWN_FAMILY_NAME
Placeholder for unknown family name 999

UNKNOWN_GIVEN_NAME
Placeholder for unknown given name 99

UNKNOWN_SEX
Placeholder for unknown sex 3

Functions

encode
This function encodes given family name, given name, date of birth and sex in XXXZZDDMMYYYYN sequence.